Revise: generate a new 1024x1024 SVG and XCF (GIMP) icon master file
The SVG file has been taken from one of the other SVG files in the
repository and unneeded elements removed, wanted elements named and grouped
and then each element exported as a 1024x1024 PNG file that has then been
imported into the GIMP as separate layers. This then leaves a GIMP native
file at a very high resolution which can subsequently be ammended to add
the dev and ptb changes to yield icons for those variants that will be
suitable for Retina/UHD/4K display systems.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>