Revise: update with modifications I have made to the splash-screens (#10)
Added layers to "working" GIMP file `Mudlet_splashscreen_working.xcf`: "brick-wall", "blackhole" and "mark 1 eyeball", and some repositioning of "logo Mudlet offset" (so it wasn't offset so much) and toggling some layers... 

Revised to import old "logo Mudlet offset" layer from prior version and renamed new one to "logo Mudlet central" so that it is still present in the file for reuse if needed. Other than toggling visibilities of other layers those should not have been changed.

The other addition being `splash_no_arch.svg` is a file modified from `splash.svg` to remove the arch and fill in with "bricks" copied and modified from the existing ones. The `.png` placed into the main repo was exported from Inkscape directly, however it proved necessary to offset the "udlet" text and that was done to the `.png` file directly so it is not exactly reproducible from this file. 

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>