Add: image for discord invite
The Discord guidance for this suggests a minimum of 1920x1080 pixels
which I have created by scaling the original `badge.svg` and adding a
rectangular area below the original rectangular wall part of logo so that
when scaled to the 16:9 ratio the wall part has the same aspect ratio as
before.

The "frame" part of the image has been turned off (hidden) otherwise it
would overlap the new rectangle - though perhaps I could have adjusted it
to completely surround the whole visible image.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>